Surface mobile_listInstalledApps to the LLM and wire up iOS
`mobile_listInstalledApps` already existed but was a hidden, callback-only
primitive (`surfaceToLlm = false`) and Android-only, so an agent driving via
the toolbox couldn't discover it:

  - `trailblaze toolbox --name/--search` filters out surfaceToLlm=false tools
    (catalog gate), so the tool was invisible.
  - `trailblaze tool mobile_listInstalledApps` failed with "not valid for the
    current device/target" because getCurrentToolDescriptors() applies the same
    surfaceToLlm filter to the available-tools set.

Make it a first-class, cross-platform tool:

  - Remove `surfaceToLlm = false` so it appears in the toolbox catalog and is
    invocable via `trailblaze tool` and the LLM. Result shape is unchanged
    (`{"appIds":[...]}`) so existing scripted-tool consumers and the CI trail
    keep working.
  - Implement the iOS branch via a new IosHostSimctlUtils.listInstalledAppIds
    (`xcrun simctl listapps`), mirroring how mobile_setClipboard reaches iOS
    from trailblaze-common. IosHostUtils.getInstalledAppIds now delegates to
    that canonical primitive (single source of truth); its parser + unit tests
    moved into trailblaze-common.
  - Anchor the listapps parser to block-opening header lines (`"<id>" = {`)
    rather than any `"key" = value` line, so nested GroupContainers entries
    (e.g. team-prefixed `"243LU…groups.com.apple.podcasts"`) aren't mis-counted
    as installed apps. Verified against real `simctl listapps` output on an
    Xcode 26.6 simulator (30 apps, no false positives).
  - Update toolset comments/descriptions, the TS SDK doc, and regenerate the
    framework reference page.

Fixes #171
