+++ trailblaze-host/src/test/java/xyz/block/trailblaze/host/ios/IosHostUtilsTest.kt
@@ -3,7 +3,6 @@ package xyz.block.trailblaze.host.ios
 import org.junit.Test
 import xyz.block.trailblaze.devices.TrailblazeDeviceId
 import xyz.block.trailblaze.devices.TrailblazeDevicePlatform
-import kotlin.test.assertContains
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
@@ -12,117 +11,9 @@ import xyz.block.trailblaze.util.Console
 
 class IosHostUtilsTest {
 
-  // region parseInstalledAppIdsFromListApps tests
-
-  @Test
-  fun `parseInstalledAppIdsFromListApps extracts bundle IDs from typical listapps output`() {
-    // Sample output from `xcrun simctl listapps <device-id>`
-    val outputLines = """
-      {
-          "com.apple.Preferences" =     {
-              ApplicationType = System;
-              Bundle = "file:///Applications/Preferences.app/";
-              CFBundleDisplayName = Settings;
-              CFBundleExecutable = Preferences;
-              CFBundleIdentifier = "com.apple.Preferences";
-              CFBundleName = Settings;
-              CFBundleVersion = 1;
-              Path = "/Applications/Preferences.app";
-          };
-          "com.example.app" =     {
-              ApplicationType = User;
-              Bundle = "file:///Users/test/Library/Developer/CoreSimulator/Devices/ABC123/data/Containers/Bundle/Application/DEF456/ExampleApp.app/";
-              CFBundleDisplayName = ExampleApp;
-              CFBundleIdentifier = "com.example.app";
-              CFBundleVersion = 6940515;
-              Path = "/Users/test/Library/Developer/CoreSimulator/Devices/ABC123/data/Containers/Bundle/Application/DEF456/ExampleApp.app";
-          };
-          "com.apple.mobilesafari" =     {
-              ApplicationType = System;
-              CFBundleIdentifier = "com.apple.mobilesafari";
-              Path = "/Applications/MobileSafari.app";
-          };
-      }
-    """.trimIndent().lines()
-
-    val appIds = IosHostUtils.parseInstalledAppIdsFromListApps(outputLines)
-
-    assertEquals(3, appIds.size)
-    assertContains(appIds, "com.apple.Preferences")
-    assertContains(appIds, "com.example.app")
-    assertContains(appIds, "com.apple.mobilesafari")
-  }
-
-  @Test
-  fun `parseInstalledAppIdsFromListApps filters out group identifiers`() {
-    val outputLines = """
-      {
-          "com.apple.Preferences" =     {
-              Path = "/Applications/Preferences.app";
-          };
-          "group.com.example.app" =     {
-              Path = "/some/group/path";
-          };
-          "group.com.example.shared" =     {
-              Path = "/another/group/path";
-          };
-          "com.example.app" =     {
-              Path = "/Applications/Example.app";
-          };
-      }
-    """.trimIndent().lines()
-
-    val appIds = IosHostUtils.parseInstalledAppIdsFromListApps(outputLines)
-
-    assertEquals(2, appIds.size)
-    assertContains(appIds, "com.apple.Preferences")
-    assertContains(appIds, "com.example.app")
-    assertTrue(appIds.none { it.startsWith("group.") })
-  }
-
-  @Test
-  fun `parseInstalledAppIdsFromListApps handles variable whitespace`() {
-    // Test with different amounts of leading whitespace and spacing around equals
-    val outputLines = listOf(
-      "{",
-      "    \"com.app.fourspaces\" =     {",      // 4 spaces, multiple spaces around =
-      "\t\"com.app.tab\" =\t{",                   // tab, tab around =
-      "  \"com.app.twospaces\" = {",              // 2 spaces, single space around =
-      "      \"com.app.sixspaces\"={",            // 6 spaces, no spaces around =
-      "}",
-    )
-
-    val appIds = IosHostUtils.parseInstalledAppIdsFromListApps(outputLines)
-
-    assertEquals(4, appIds.size)
-    assertContains(appIds, "com.app.fourspaces")
-    assertContains(appIds, "com.app.tab")
-    assertContains(appIds, "com.app.twospaces")
-    assertContains(appIds, "com.app.sixspaces")
-  }
-
-  @Test
-  fun `parseInstalledAppIdsFromListApps returns empty set for empty input`() {
-    val appIds = IosHostUtils.parseInstalledAppIdsFromListApps(emptyList())
-    assertTrue(appIds.isEmpty())
-  }
-
-  @Test
-  fun `parseInstalledAppIdsFromListApps ignores lines without app identifiers`() {
-    val outputLines = listOf(
-      "{",
-      "    ApplicationType = System;",
-      "    CFBundleDisplayName = Settings;",
-      "    Path = \"/Applications/Preferences.app\";",
-      "};",
-      "}",
-    )
-
-    val appIds = IosHostUtils.parseInstalledAppIdsFromListApps(outputLines)
-    assertTrue(appIds.isEmpty())
-  }
-
-  // endregion
+  // NOTE: `parseInstalledAppIdsFromListApps` moved to `IosHostSimctlUtils` (trailblaze-common) so
+  // the cross-platform `mobile_listInstalledApps` tool can share one source of truth. Its unit
+  // tests now live in `IosHostSimctlUtilsTest`.
 
   // region parseAppPathFromListApps tests
