+++ trailblaze-models/src/commonMain/resources/trails/config/trailmaps/android/toolsets/android_framework.yaml
@@ -4,16 +4,19 @@
 # TS handlers call `mobile_listInstalledApps` to resolve the installed Square app id, then
 # `android_sendBroadcast` to fire the actual intent.
 #
-# Each tool in here is declared `surfaceToLlm = false` at the class level, so this toolset does
-# NOT pollute the LLM's tool catalog. The point of being a toolset (rather than discoverable
-# only via `ToolYamlLoader`) is to land these tools in the runtime `TrailblazeToolRepo` so
-# the daemon's `/scripting/callback` dispatcher can resolve them by name. Without this
-# toolset, the callback handler throws "Could not find Trailblaze tool for name: …" the
-# moment a scripted tool composes one of them.
+# `android_sendBroadcast` / `android_adbShell` are declared `surfaceToLlm = false` at the class
+# level, so they do NOT pollute the LLM's tool catalog. `mobile_listInstalledApps` is LLM-facing
+# (so an agent can discover installed apps before `launchApp`) but is grouped here because the same
+# Square card-reader handlers compose it via the callback channel. The point of being a toolset
+# (rather than discoverable only via `ToolYamlLoader`) is to land these tools in the runtime
+# `TrailblazeToolRepo` so the daemon's `/scripting/callback` dispatcher can resolve them by name.
+# Without this toolset, the callback handler throws "Could not find Trailblaze tool for name: …"
+# the moment a scripted tool composes one of them.
 #
-# `always_enabled: true` is load-bearing — the LLM never calls `setActiveToolSets` for these
-# (it doesn't see them in the first place), so they have to be active from session start or
-# the very first callback fails.
+# `always_enabled: true` is load-bearing — for the two `surfaceToLlm = false` tools the LLM never
+# calls `setActiveToolSets` (it doesn't see them in the first place), so they have to be active
+# from session start or the very first callback fails. The LLM-facing `mobile_listInstalledApps`
+# piggy-backs on the same flag.
 #
 # Driver scope matches where these primitives have working on-device actuals today (host-
 # driven Android sessions reach the device through these same on-device drivers via RPC, so
@@ -21,7 +24,7 @@
 # each tool throws a "not yet supported on iOS" error, so registering them there would only
 # surface as confusing runtime failures rather than catalog-time filtering.
 id: android_framework
-description: "Framework-level Android composition primitives used by scripted tools via `client.callTool(...)`. Not LLM-facing — registered so callback dispatch can resolve them by name."
+description: "Framework-level Android composition primitives used by scripted tools via `client.callTool(...)`. `android_sendBroadcast` / `android_adbShell` are not LLM-facing (registered so callback dispatch can resolve them by name); `mobile_listInstalledApps` is LLM-facing so an agent can discover installed apps."
 always_enabled: true
 drivers:
   - android-ondevice-instrumentation
