+++ trailblaze-models/src/commonMain/resources/trails/config/trailmaps/mobile/toolsets/mobile_primitives.yaml
@@ -3,33 +3,32 @@
 # carried the flat name prefix; this toolset makes the prefix-equals-trailmap-id rule hold by
 # grouping them under a trailmap-aligned name.
 #
-# Mixed LLM surfacing — `mobile_pasteClipboard` / `mobile_setClipboard` are LLM-facing
-# (recordable interaction primitives), while `mobile_clearAppData` / `mobile_listInstalledApps`
-# are declared `surfaceToLlm = false` (callback-only composition primitives). The toolset
-# groups them together because they share the same `mobile_*` prefix and live behind the same
-# library trailmap; the per-tool `surfaceToLlm` flag still controls catalog visibility, so a
-# consumer that pulls in this toolset surfaces only the two LLM-facing entries to its agent
-# while still registering all four for `client.callTool(...)` dispatch.
+# Mixed LLM surfacing — `mobile_pasteClipboard` / `mobile_setClipboard` / `mobile_listInstalledApps`
+# are LLM-facing (the clipboard tools are recordable interaction primitives; `listInstalledApps`
+# lets an agent discover installed apps before `launchApp`), while `mobile_clearAppData` is declared
+# `surfaceToLlm = false` (a callback-only composition primitive). The toolset groups them together
+# because they share the same `mobile_*` prefix and live behind the same library trailmap; the
+# per-tool `surfaceToLlm` flag still controls catalog visibility, so a consumer that pulls in this
+# toolset surfaces only the LLM-facing entries to its agent while still registering all four for
+# `client.callTool(...)` dispatch.
 #
-# `always_enabled: true` is load-bearing for the two `surfaceToLlm = false` tools — they're
-# invisible to `setActiveToolSets`, so they have to be active from session start or the first
-# callback dispatching `mobile_clearAppData` / `mobile_listInstalledApps` fails. The LLM-facing
-# entries piggy-back on the same flag, matching how `core_interaction` (which currently owns
-# the clipboard tools) is also `always_enabled: true`.
+# `always_enabled: true` is load-bearing for `mobile_clearAppData` (`surfaceToLlm = false`) — it's
+# invisible to `setActiveToolSets`, so it has to be active from session start or the first callback
+# dispatching it fails. The LLM-facing entries piggy-back on the same flag, matching how
+# `core_interaction` (which currently owns the clipboard tools) is also `always_enabled: true`.
 #
 # Driver scope mirrors where each primitive actually has a working on-device path today:
 #  - Android on-device drivers cover all four primitives (clearAppData / listInstalledApps via
 #    on-device executors, paste/setClipboard via UiAutomator).
-#  - `ios-host` covers `mobile_clearAppData` (via `xcrun simctl`) and the clipboard tools
-#    (the iOS clipboard accessibility-label path). `mobile_listInstalledApps` has no working
-#    iOS implementation today and would surface as a runtime error if invoked on iOS — that's
-#    a tool-level concern, not a toolset-scope concern.
+#  - `ios-host` covers `mobile_clearAppData` (via `xcrun simctl`), the clipboard tools (the iOS
+#    clipboard accessibility-label path), and `mobile_listInstalledApps` (via `xcrun simctl
+#    listapps`).
 #
 # NOTE: this toolset is additive. The existing `core_interaction` (paste/setClipboard) and
 # `android_framework` (listInstalledApps) assignments are intentionally untouched in this
 # PR — future PRs can migrate consumers once the downstream migration is ready. See PR #3435.
 id: mobile_primitives
-description: "Cross-platform mobile composition primitives (clearAppData, listInstalledApps, paste/setClipboard). LLM-facing entries (paste/setClipboard) appear in the agent catalog; the surfaceToLlm=false entries (clearAppData, listInstalledApps) are dispatch-only via `client.callTool(...)`."
+description: "Cross-platform mobile composition primitives (clearAppData, listInstalledApps, paste/setClipboard). LLM-facing entries (paste/setClipboard, listInstalledApps) appear in the agent catalog; the surfaceToLlm=false entry (clearAppData) is dispatch-only via `client.callTool(...)`."
 always_enabled: true
 drivers:
   - android-ondevice-instrumentation
